Refactored Deepbots: DeepbotsEnv & Robot Class
1. Inherited from Webots Robot class
2. Refactored SupervisorEnv name to DeepbotsEnv
3. Shortened imports
4. Removed setup folder
5. Renamed lots of class name
