+++ deepbots/robots/__init__.py
@@ -1,0 +1,2 @@
+from deepbots.robots.controllers.csv_robot import CSVRobot
+from deepbots.robots.controllers.emitter_receiver_robot import EmitterReceiverRobot
