+++ deepbots/supervisor/__init__.py
@@ -1,0 +1,4 @@
+from deepbots.supervisor.controllers.robot_supervisor_env import RobotSupervisorEnv
+from deepbots.supervisor.controllers.csv_supervisor_env import CSVSupervisorEnv
+from deepbots.supervisor.controllers.deepbots_supervisor_env import DeepbotsSupervisorEnv
+from deepbots.supervisor.controllers.emitter_receiver_supervisor_env import EmitterReceiverSupervisorEnv
