+++ deepbots/supervisor/controllers/emitter_receiver_supervisor_env.py
@@ -1,11 +1,10 @@
-from collections.abc import Iterable
 from warnings import warn, simplefilter
 
-from deepbots.supervisor.controllers.supervisor_env import SupervisorEnv
+from deepbots.supervisor.controllers.deepbots_supervisor_env import DeepbotsSupervisorEnv
 from controller import Supervisor
 
 
-class SupervisorEmitterReceiver(SupervisorEnv):
+class EmitterReceiverSupervisorEnv(DeepbotsSupervisorEnv):
     """
     This is the base class for the emitter - receiver scheme.
 
@@ -26,7 +25,7 @@ class SupervisorEmitterReceiver(SupervisorEnv):
             supervisor node
         :param timestep: The supervisor controller timestep
         """
-        super(SupervisorEmitterReceiver, self).__init__()
+        super(EmitterReceiverSupervisorEnv, self).__init__()
 
         if timestep is None:
             self.timestep = int(self.getBasicTimeStep())
@@ -116,61 +115,3 @@ class SupervisorEmitterReceiver(SupervisorEnv):
         :param value: The new controller timestep in milliseconds
         """
         self._timestep = int(value)
-
-
-class SupervisorCSV(SupervisorEmitterReceiver):
-    """
-    This class implements the emitter-receiver scheme using Comma Separated
-    Values.
-    """
-    def __init__(self,
-                 emitter_name="emitter",
-                 receiver_name="receiver",
-                 timestep=None):
-        """
-        The constructor just passes the arguments provided to the parent
-        class contructor.
-
-        :param emitter_name: The name of the emitter device on the
-            supervisor node
-        :param receiver_name: The name of the receiver device on the
-            supervisor node
-        :param timestep: The supervisor controller timestep
-        """
-        super(SupervisorCSV, self).__init__(emitter_name, receiver_name,
-                                            timestep)
-
-    def handle_emitter(self, action):
-        """
-        Implementation of the handle_emitter method expecting an iterable
-        with Comma Separated Values (CSV).
-
-        :param action: Whatever the use-case uses as an action, e.g.
-            an integer representing discrete actions
-        :type action:  Iterable, for multiple values the CSV format is
-            required, e.g. [0, 1] for two actions
-        """
-        assert isinstance(action, Iterable), \
-            "The action object should be Iterable"
-
-        message = (",".join(map(str, action))).encode("utf-8")
-        self.emitter.send(message)
-
-    def handle_receiver(self):
-        """
-        Implementation of the handle_receiver method expecting an iterable
-        with Comma Separated Values (CSV).
-
-        :return: Returns the message received from the robot, returns None
-            if no message is received
-        :rtype: List of string values
-        """
-        if self.receiver.getQueueLength() > 0:
-            try:
-                string_message = self.receiver.getString()
-            except AttributeError:
-                string_message = self.receiver.getData().decode("utf-8")
-            self.receiver.nextPacket()
-            return string_message.split(",")
-        else:
-            return None
