+++ deepbots/supervisor/wrappers/__init__.py
@@ -1,0 +1,2 @@
+from deepbots.supervisor.wrappers.keyboard_printer import KeyboardPrinter
+from deepbots.supervisor.wrappers.tensorboard_wrapper import TensorboardLogger
