+++ deepbots/supervisor/wrappers/keyboard_printer.py
@@ -1,16 +1,16 @@
 from controller import Keyboard
 
-from deepbots.supervisor.controllers.supervisor_env import SupervisorEnv
+from deepbots.supervisor.controllers.deepbots_supervisor_env import DeepbotsSupervisorEnv
 
 
-class KeyboardPrinter(SupervisorEnv):
+class KeyboardPrinter(DeepbotsSupervisorEnv):
     def __init__(self, controller):
         self.controller = controller
         self.keyboard = Keyboard()
         self.keyboard.enable(self.controller.timestep)
 
     def step(self, action):
-        observation, reward, isDone, info = self.controller.step(action)
+        observation, reward, is_done, info = self.controller.step(action)
         key = self.keyboard.getKey()
         # DEBUG CONTROLS
         if key == Keyboard.CONTROL + ord("A"):
@@ -23,13 +23,13 @@ class KeyboardPrinter(SupervisorEnv):
             print()
             print("Observations: ", self.controller.observation)
 
-        return observation, reward, isDone, info
+        return observation, reward, is_done, info
 
     def is_done(self):
-        isDone = self.controller.is_done()
-        if isDone:
+        is_done = self.controller.is_done()
+        if is_done:
             print("Done")
-        return isDone
+        return is_done
 
     def get_observations(self):
         return self.controller.get_observations()
